lab_7/v1/fastapi-nginx-app_v2.0.2/app/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python

